# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Swagger Mono-Api"
    },
    "host": "localhost:8081",
    "basePath": "/techmillresource/mono-api/api",
...
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "uikey": {
            "in": "header",
            "name": "uikey",
            "type": "apiKey"
        },
        "session_key": {
            "in": "header",
            "name": "apisessionkey",
            "type": "apiKey"
        }
    },
    "paths": {
...
        "/getTransaction": {
            "post": {
                "tags": [
                    "transaction"
                ],
                "description": "user(s) transaction details in system",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "transaction details of user in system",
                    "schema": {
                        "$ref": "#/definitions/GetTransaction"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/GetTransaction"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [{
                    "uikey": [],
                    "session_key": []
                }]
            }
        },
        "/getReferenceNo": {
            "post": {
                "tags": [
                    "transaction"
                ],
                "description": "generate a reference number for user(s) in system",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "reference number generated for user in system",
                    "schema": {
                        "$ref": "#/definitions/GetReferenceNo"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/GetReferenceNo"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [{
                    "uikey": [],
                    "session_key": []
                }]
            }
        },
        "/createTransaction": {
            "post": {
                "tags": [
                    "transaction"
                ],
                "description": "user transaction details to create in system",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "transaction details of user to create in system",
                    "schema": {
                        "$ref": "#/definitions/CreateTransaction"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/CreateTransaction"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [{
                    "uikey": [],
                    "session_key": []
                }]
            }
        }
    },
    "definitions": {
...
        "GetTransaction": {
            "type": "object",
            "required": [
                "profileId",
                "channel"
            ],
            "properties": {
                "profileId": {
                    "type": "string",
                    "uniqueItems": true
                },
                "channel": {
                    "type": "string",
                    "uniqueItems": true,
                    "enum": [
                        "mobile",
                        "self service"
                    ]
                }
            }
        },
        "GetReferenceNo": {
            "type": "object",
            "required": [
                "profileId",
                "channel"
            ],
            "properties": {
                "profileId": {
                    "type": "string",
                    "uniqueItems": true
                },
                "channel": {
                    "type": "string",
                    "uniqueItems": true,
                    "enum": [
                        "mobile",
                        "self service"
                    ]
                }
            }
        },
        "CreateTransaction": {
            "type": "object",
            "required": [
                "profileId",
                "transactionReference",
                "transactionDescription",
                "transactionFee",
                "transactionAmount",
                "channel"
            ],
            "properties": {
                "profileId": {
                    "type": "string",
                    "uniqueItems": true
                },
                "transactionReference": {
                    "type": "string",
                    "uniqueItems": true
                },
                "transactionDescription": {
                    "type": "string",
                    "uniqueItems": true
                },
                "transactionFee": {
                    "type": "number",
                    "uniqueItems": true
                },
                "transactionAmount": {
                    "type": "number",
                    "uniqueItems": true
                },
                "channel": {
                    "type": "string",
                    "uniqueItems": true,
                    "enum": [
                        "mobile",
                        "self service"
                    ]
                }
            }
        }
    }
}